+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 - [TD-7539] Allow to download notes from ElasticSearch for structure details
 - [TD-7396] Adjusts implementation actions permissions for users with permissions in different domains
+- [TD-7340] Support for `dynamic_table` template fields
 
 ## [7.11.2] 2025-10-27